Split the tests into multiple files
so that protractor flake can re-run only that test file and not all of the tests.
