+++ package.json
@@ -49,7 +49,7 @@
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^1.7.0",
     "opentok-test-scripts": "^3.0.3",
-    "protractor": "3.2.2",
+    "protractor": "^3.2.2",
     "protractor-flake": "^2.1.2",
     "q": "^1.2.0",
     "travis-multirunner": "^3.0.0",
