+++ tests/protractor.conf.js
@@ -25,7 +25,7 @@ switch(process.env.BROWSER) {
       sauceKey: process.env.SAUCE_ACCESS_KEY,
 
       specs: [
-        'e2e/iesmoketest.js'
+        'e2e/ie/iesmoketest.js'
       ],
 
       capabilities: getCapabilitiesFor(process.env.BROWSER, process.env.BVER),
@@ -46,7 +46,7 @@ switch(process.env.BROWSER) {
       allScriptsTimeout: 11000,
 
       specs: [
-        'e2e/scenarios.js'
+        'e2e/*.js'
       ],
 
       getMultiCapabilities: helper.getFirefoxProfile,
@@ -77,7 +77,7 @@ switch(process.env.BROWSER) {
       allScriptsTimeout: 11000,
 
       specs: [
-        'e2e/scenarios.js'
+        'e2e/*.js'
       ],
 
       capabilities: {
